chrome/test/data/xr/e2e_test_files/html/webxr_test_presentation_promise_rejected_if_consent_not_granted.html

<!doctype html>
<!--
Used to test that the promise returned by WebXR's requestSession rejects if
the user denies permission on a consent dialog.
-->
<html>
  <head>
    <link rel="stylesheet" type="text/css" href="../resources/webxr_e2e.css">
  </head>
  <body>
    <canvas id="webgl-canvas"></canvas>
    <script src="../../../../../../third_party/blink/web_tests/resources/testharness.js"></script>
    <script src="../resources/webxr_e2e.js"></script>
    <script src="../resources/webxr_boilerplate.js"></script>
    <script>
      function onImmersiveRequestWithConsent() {
        navigator.xr.requestSession('immersive-vr').then(() => {
          assert_unreached("requestPresent promise shouldn't resolve when consent not granted");
        }).catch((e) => {
          assert_true(e instanceof DOMException && e.name == "NotSupportedError");
          done();
        });
      }
    </script>
  </body>
</html>
